refactor(GoalService): add authenticated user

+++ src/main/java/com/pomostudy/service/GoalService.java
@@ -1,5 +1,7 @@
 package com.pomostudy.service;
 
+import com.pomostudy.config.security.AuthenticatedUser;
+import com.pomostudy.entity.Task;
 import com.pomostudy.exception.ResourceExceptionFactory;
 import com.pomostudy.mapper.GoalMapper;
 import com.pomostudy.dto.goal.GoalRequestDTO;
@@ -16,41 +18,79 @@ import java.util.Optional;
 @Service
 public class GoalService {
 
-    final GoalRepository goalRepository;
-    final UserRepository userRepository;
-    final GoalMapper goalMapper;
+    private final GoalRepository goalRepository;
+    private final UserRepository userRepository;
+    private final GoalMapper goalMapper;
+    private final AuthorizationService authorizationService;
 
-    public GoalService(GoalRepository goalRepository, UserRepository userRepository, GoalMapper goalMapper) {
+    private static final String GOAL = "Goal";
+
+    public GoalService(
+            GoalRepository goalRepository,
+            UserRepository userRepository,
+            GoalMapper goalMapper,
+            AuthorizationService authorizationService
+
+    ) {
         this.goalRepository = goalRepository;
         this.userRepository = userRepository;
         this.goalMapper = goalMapper;
+        this.authorizationService = authorizationService;
     }
 
-    public GoalResponseDTO save(GoalRequestDTO goalRequestDTO) {
-        Goal goal = goalRepository.save(goalMapper.toCreateGoal(goalRequestDTO));
+    public GoalResponseDTO save(GoalRequestDTO goalRequestDTO, AuthenticatedUser authenticatedUser) {
+        Goal goal = goalRepository.save(goalMapper.toCreateGoal(goalRequestDTO, authenticatedUser));
         return goalMapper.toGoalResponseDTO(goal);
     }
 
-    public GoalResponseDTO edit(GoalRequestDTO goalRequestDTO, Long id) {
-        Goal goal = goalRepository.save(goalMapper.toUpdateGoal(goalRequestDTO, id));
+    public GoalResponseDTO edit(GoalRequestDTO goalRequestDTO, AuthenticatedUser authenticatedUser, Long id) {
+
+        Long userId = authenticatedUser.getUser().getId();
+
+        if (!authorizationService.isOwner(Goal.class, id, userId) && !authenticatedUser.isAdmin()) {
+            throw ResourceExceptionFactory.notFound(GOAL, id);
+        }
+
+        Goal goal = goalRepository.save(goalMapper.toUpdateGoal(goalRequestDTO, authenticatedUser, id));
         return goalMapper.toGoalResponseDTO(goal);
 
     }
 
-    public Page<GoalResponseDTO> findAll(Pageable pageable) {
-        Page<Goal> goalPage = goalRepository.findAll(pageable);
+    public Page<GoalResponseDTO> findAll(Pageable pageable, AuthenticatedUser authenticatedUser) {
+        Page<Goal> goalPage;
+
+        if (authenticatedUser.isAdmin()) {
+            goalPage = goalRepository.findAll(pageable);
+        } else {
+            goalPage = goalRepository.findByUser(authenticatedUser.getUser(), pageable);
+        }
+
         return goalPage.map(goalMapper::toGoalResponseDTO);
     }
 
-    public GoalResponseDTO findById(Long id) {
+    public GoalResponseDTO findById(Long id, AuthenticatedUser authenticatedUser) {
+
+        Long userId = authenticatedUser.getUser().getId();
+
+        if (!authorizationService.isOwner(Goal.class, id, userId) && !authenticatedUser.isAdmin()) {
+            throw ResourceExceptionFactory.notFound(GOAL, id);
+        }
+
         return goalRepository.findById(id)
                 .map(GoalResponseDTO::new)
-                .orElseThrow(() -> ResourceExceptionFactory.notFound("Goal", id));
+                .orElseThrow(() -> ResourceExceptionFactory.notFound(GOAL, id));
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, AuthenticatedUser authenticatedUser) {
+
+        Long userId = authenticatedUser.getUser().getId();
+
+        if (!authorizationService.isOwner(Goal.class, id, userId) && !authenticatedUser.isAdmin()) {
+            throw ResourceExceptionFactory.notFound(GOAL, id);
+        }
+
         goalRepository.findById(id)
-                .orElseThrow(() -> ResourceExceptionFactory.notFound("Goal", id));
+                .orElseThrow(() -> ResourceExceptionFactory.notFound(GOAL, id));
         goalRepository.deleteById(id);
     }
 }
